Add files via upload
Erratum: In the paper titled "Non-destructive estimation of wood-boring pest density in living trees using X-ray imaging and Edge Computing techniques," an evident detail error was identified in Fig. 12. The authors have made the necessary corrections, and this error does not affect the overall content or findings of the paper.